Add files via upload
This dataset consists of the marks secured by the students in various subjects. The hope is to understand the influence of the parents’ background, test preparation etc. on students performance. Will try to look at each variable and also their relationships with creating a detailed statistical analysis of the data through both R script and graphs. Then will be finalized applying Linear Regression.